Tweak flash interface
On the implementation but not the types I've left open the possibility
of passing a MemoryMap so I can experiement with it as an option.

This means we have a hard dependency on the intel hex library but I
think we should either stick with that or have a slimline version.