Tweak flash interface (#7)
This also fixes a mistake where we padded with 0 not 0xff.

On the implementation but not the types I've left open the possibility of passing a MemoryMap so I can experiment with it as an option.

MemoryMap imports remain a bit of a disaster but one to investigate further another day.

For the Python Editor this will mean that we always return a hex from the file system and then parse it for the partial flashing case. But that seems a much more reasonable API and the conversion should be quick enough. If we care for perf then we can work towards using the MemoryMap built by microbit-fs directly. There's no reason to be going via padded bytes as that drops the UICR data which it would be better to flash.