Add deviceSelectionMode to MicrobitWebUSBConnectionOptions (#60)
Added a new configuration (`deviceSelectionMode`) for how a device should be selected.

`DeviceSelectionMode.AlwaysAsk` - Attempts to connect to known device, otherwise asks which device to connect with.
`DeviceSelectionMode.UseAnyAllowed` - Attempts to connect to known device, otherwise attempts to connect with any allowed devices. If that fails, asks which device to connect with.